Split up Node Graph across multiple files
