+++ src/gui/models/CMakeLists.txt
@@ -1,3 +1,5 @@
+add_subdirectory(nodeGraph)
+
 set(models_MOC_HDRS
     addForcefieldDialogModel.h
     addSpeciesInfoModel.h
@@ -10,7 +12,6 @@ set(models_MOC_HDRS
     dissolveModel.h
     enumOptionsModel.h
     forcefieldModel.h
-    graphModel.h
     isotopologueSetModel.h
     masterAngleModel.h
     masterBondModel.h
@@ -74,7 +75,6 @@ set(models_SRCS
     ffSortFilterModel.cpp
     forcefieldModel.cpp
     globalPotentialFilterProxy.cpp
-    graphModel.cpp
     isotopologueSetModel.cpp
     masterAngleModel.cpp
     masterBondModel.cpp
@@ -162,7 +162,7 @@ target_include_directories(
                  ${FREETYPE_INCLUDE_DIRS} ${Qt6Quick_INCLUDE_DIRS}
 )
 
-target_link_libraries(models PRIVATE Qt6::Core Qt6::Qml)
+target_link_libraries(models PRIVATE Qt6::Core Qt6::Qml nodeGraphModel)
 
 # Because Qt has its own way of checking for C++17
 if(MSVC)
