+++ src/gui/models/types.cpp
@@ -7,7 +7,7 @@
 #include "gui/models/configurationModel.h"
 #include "gui/models/dataManagerSimulationModel.h"
 #include "gui/models/dissolveModel.h"
-#include "gui/models/graphModel.h"
+#include "gui/models/nodeGraph/graphModel.h"
 #include "gui/models/masterTermTreeModel.h"
 #include "gui/models/modifyChargesModel.h"
 #include "gui/models/moduleLayersModel.h"
